src/requirements/core.txt
fastapi==0.103.1
simpy==4.0.2
orjson==3.9.5
flower==2.0.1
httpx==0.24.1
celery==5.2.3
redis==4.3.4
uvicorn==0.23.2
pydantic-settings==2.0.3
pika==1.3.2
PyYAML
